Add Mustache Express template engine and handler
We've added the 'mustache-express' library as a dependency, and also implemented template rendering using Mustache for an endpoint in the UserController. The setup for using Mustache Express as the application's view engine was performed in 'main.ts'. A new page, 'index.html', was added in our views directory with a simple greeting template.