Implement MemberService and integrate it into UserController
Created a new MemberService class to manage Member related operations such as getting connection name and sending email. The service was also integrated into the UserController with required calls made to the newly implemented methods. In addition, MemberService was added to the providers in the UserModule. Corresponding test spec created for MemberService as well.